Fix some uint16 errors

+++ converted/sounds.go
@@ -61,7 +61,7 @@ func SoundInitFreqTable() {
 	for octave = 1; octave <= 15; octave++ {
 		noteBase = Exp(float64(octave)*ln2) * freqC1
 		for note = 0; note <= 11; note++ {
-			SoundFreqTable[(octave*16+note)-1] = Trunc(noteBase)
+			SoundFreqTable[(octave*16+note)-1] = uint16(Trunc(noteBase))
 			noteBase = noteBase * noteStep
 		}
 	}
@@ -75,29 +75,29 @@ func SoundInitDrumTable() {
 		SoundDrumTable[i].Len = 14
 	}
 	for i = 1; i <= 14; i++ {
-		SoundDrumTable[1].Data[i-1] = i*100 + 1000
+		SoundDrumTable[1].Data[i-1] = uint16(i*100 + 1000)
 	}
 	for i = 1; i <= 16; i++ {
-		SoundDrumTable[2].Data[i-1] = (i%2)*1600 + 1600 + (i%4)*1600
+		SoundDrumTable[2].Data[i-1] = uint16((i%2)*1600 + 1600 + (i%4)*1600)
 	}
 	for i = 1; i <= 14; i++ {
-		SoundDrumTable[4].Data[i-1] = Random(5000) + 500
+		SoundDrumTable[4].Data[i-1] = uint16(Random(5000) + 500)
 	}
 	for i = 1; i <= 8; i++ {
 		SoundDrumTable[5].Data[(i*2-1)-1] = 1600
-		SoundDrumTable[5].Data[(i*2)-1] = Random(1600) + 800
+		SoundDrumTable[5].Data[(i*2)-1] = uint16(Random(1600) + 800)
 	}
 	for i = 1; i <= 14; i++ {
-		SoundDrumTable[6].Data[i-1] = ((i % 2) * 880) + 880 + ((i % 3) * 440)
+		SoundDrumTable[6].Data[i-1] = uint16(((i % 2) * 880) + 880 + ((i % 3) * 440))
 	}
 	for i = 1; i <= 14; i++ {
-		SoundDrumTable[7].Data[i-1] = 700 - (i * 12)
+		SoundDrumTable[7].Data[i-1] = uint16(700 - (i * 12))
 	}
 	for i = 1; i <= 14; i++ {
-		SoundDrumTable[8].Data[i-1] = (i*20 + 1200) - Random(i*40)
+		SoundDrumTable[8].Data[i-1] = uint16((i*20 + 1200) - Random(i*40))
 	}
 	for i = 1; i <= 14; i++ {
-		SoundDrumTable[9].Data[i-1] = Random(440) + 220
+		SoundDrumTable[9].Data[i-1] = uint16(Random(440) + 220)
 	}
 }
 
@@ -136,7 +136,7 @@ func SoundHasTimeElapsed(counter *int16, duration int16) (SoundHasTimeElapsed bo
 		hSecsDiff = Word((hSecsTotal-*counter)+6000) % 6000
 	} else {
 		hSecsTotal = int16(TimerTicks * 6)
-		hSecsDiff = hSecsTotal - *counter
+		hSecsDiff = uint16(hSecsTotal - *counter)
 	}
 	if hSecsDiff >= uint16(duration) {
 		SoundHasTimeElapsed = true
